# openapi_billing_service.yaml
openapi: 3.0.3
info:
  title: Billing Service API
  version: 1.0.0
  description: >
    Billing Service for Hospital Management System.
    Supports bill listing, bill retrieval, bill generation from completed appointments,
    cancellation handling, and billing adjustments.

servers:
  - url: http://localhost:9005

paths:
  /health:
    get:
      summary: Health check
      tags:
        - Health
      responses:
        '200':
          description: Service health status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: billing

  /v1/bills:
    get:
      summary: List bills
      tags:
        - Bills
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: patient_id
          schema:
            type: integer
        - in: query
          name: appointment_id
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            example: OPEN
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BillResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /v1/bills/{bill_id}:
    get:
      summary: Get bill by ID
      tags:
        - Bills
      parameters:
        - in: path
          name: bill_id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bill details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /v1/bills/generate-from-appointment/{appointment_id}:
    post:
      summary: Generate bill for completed appointment
      tags:
        - Bills
      parameters:
        - in: path
          name: appointment_id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bill generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillResponse'
        '400':
          description: Appointment is not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Bill already exists for appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/bills/handle-cancellation/{appointment_id}:
    post:
      summary: Handle bill update for cancelled appointment
      tags:
        - Bills
      parameters:
        - in: path
          name: appointment_id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: Bill updated after cancellation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillResponse'
        '400':
          description: Invalid cancellation request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /v1/bills/{bill_id}/adjust:
    post:
      summary: Adjust an existing bill
      tags:
        - Bills
      parameters:
        - in: path
          name: bill_id
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustmentRequest'
      responses:
        '200':
          description: Bill adjusted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillResponse'
        '400':
          description: Invalid adjustment or paid bill cannot be edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: MockToken

  schemas:
    BillResponse:
      type: object
      properties:
        bill_id:
          type: integer
          example: 1
        patient_id:
          type: integer
          example: 101
        appointment_id:
          type: integer
          example: 5001
        base_amount:
          type: number
          format: float
          example: 120.0
        tax_amount:
          type: number
          format: float
          example: 6.0
        total_amount:
          type: number
          format: float
          example: 126.0
        status:
          type: string
          example: OPEN
          enum:
            - OPEN
            - PAID
            - VOID
            - REFUND
            - PARTIAL_REFUND
        created_at:
          type: string
          format: date-time
          example: "2026-04-21T10:30:00Z"

    CancelRequest:
      type: object
      required:
        - penalty
      properties:
        penalty:
          type: number
          format: float
          description: Cancellation penalty ratio between 0 and 1
          example: 0.5

    AdjustmentRequest:
      type: object
      required:
        - adjustment_amount
        - reason
      properties:
        adjustment_amount:
          type: number
          format: float
          example: 25.0
        reason:
          type: string
          example: Additional medication charge

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "404"
        message:
          type: string
          example: Bill not found
        correlationId:
          type: string
          example: 8d6b5d4a-b0c3-4f22-b0fd-2b9c9f947f56

  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'